requirements.txt
dnspython==2.4.2
python-whois==0.8.0
colorama==0.4.6 